// src/types/types.ts
export interface Movie {
  imdbID: string;
  Title: string;
  Year: string;
  Rated: string;
  Released: string;
  Runtime: string;
  Genre: string;
  Director: string;
  Writer: string;
  Actors: string;
  Plot: string;
  Language: string;
  Country: string;
  Awards: string;
  Poster: string;
  Ratings: {
    Source: string;
    Value: string;
  }[];
  Metascore: string;
  imdbRating: string;
  imdbVotes: string;
  Type: string;
  DVD: string;
  BoxOffice: string;
  Production: string;
  Website: string;
  Response: string;
}

export interface MoviesState {
  loading: boolean;
  error: string | null;
  movies: Movie[];
  movieDetails: Movie | null;
  favorites: Movie[];
}

export interface LocationState {
  movie: {
    id: number;
    title: string;
    description: string;
    imageUrl: string;
  };
}

export interface Props {
  movie: Movie;
}